Added references to other computational studies
The Zhang 2016 paper has a bunch more of the same guests, but they were all treated as neutral, whereas the experiment had the guests charged.  I've omitted them to avoid confusion.

Also, I just realized that names like "4-methylphenylacetate" may have caused confusion in the literature.  Rekharsky 1997 used that name to refer to the charged version of 4-methylphenylacetic acid, but both Wickstrom 2013 or Zhang 2016 interpreted it as the structure that is more commonly known as "p-tolyl acetate" which is totally different.  This occurred for several different related compounds.  I've left those off the list as well, because clearly the numbers won't be comparable.